refactor(frontend): Move Tnetwork token UI to specific folder
