+++ src/frontend/src/lib/components/tokens/TokensDisplayHandler.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { nonNullish } from '@dfinity/utils';
 	import { type Snippet, untrack } from 'svelte';
-	import { sortedFungibleNetworkTokensUi } from '$lib/derived/network-tokens.derived';
+	import { sortedFungibleNetworkTokensUi } from '$lib/derived/network-tokens-ui.derived';
 	import { showZeroBalances } from '$lib/derived/settings.derived';
 	import type { TokenUiOrGroupUi } from '$lib/types/token-ui-group';
 	import { filterTokenGroups, groupTokensByTwin } from '$lib/utils/token-group.utils';
