+++ src/frontend/src/tests/lib/components/tokens/TokensDisplayHandler.spec.ts
@@ -1,4 +1,4 @@
-import { sortedFungibleNetworkTokensUi } from '$lib/derived/network-tokens.derived';
+import { sortedFungibleNetworkTokensUi } from '$lib/derived/network-tokens-ui.derived';
 import { showZeroBalances } from '$lib/derived/settings.derived';
 import { userProfileStore } from '$lib/stores/user-profile.store';
 import type { TokenUiOrGroupUi } from '$lib/types/token-ui-group';
